refactor: rename folder with middlewares, correct imports of middleware functions
