+++ app.js
@@ -8,8 +8,7 @@ const cookieParser = require('cookie-parser');
 const { PORT, BD_URL } = require('./configs');
 const limiter = require('./configs/limiter-config');
 const router = require('./routes');
-const handleError = require('./middlewares.js/handleError');
-const { requestLogger, errorLogger } = require('./middlewares.js');
+const { handleError, requestLogger, errorLogger } = require('./middlewares');
 
 const app = express();
 mongoose.connect(BD_URL, {
