+++ routes/index.js
@@ -4,7 +4,7 @@ const usersRouter = require('../components/user/router');
 const moviesRouter = require('../components/movie/router');
 const notFoundRouter = require('../components/not-found/router');
 const { createUser, signIn, signOut } = require('../components/user/controller');
-const auth = require('../middlewares.js/auth');
+const { auth } = require('../middlewares');
 
 router.post('/signup', createUser);
 router.post('/signin', signIn);
